+++ AddOns/Cell/Locales/enUS.lua
@@ -89,6 +89,50 @@ select(2, ...).L = setmetatable({
         <h2>If there are any issues after an update, check through all code snippets first.</h2>
         <br/>
 
+        <h1>r274-release (2026-01-22 19:07 GMT+8)</h1>
+        <p>* Update Molten Core debuffs (thanks Rurutia).</p>
+        <p>* Fixed an issue with getting hostile boss frames via LibGetFrame.</p>
+        <br/>
+
+        <h1>r273-release (2025-12-29 17:40 GMT+8)</h1>
+        <p>+ Added support for TBC classic.</p>
+        <p>* Improved range check.</p>
+        <p>* Updated locales.</p>
+        <br/>
+
+        <h1>r272-release (2025-12-10 11:00 GMT+8)</h1>
+        <p>* Updated debuffs for Throne of Thunder (#436, thanks Ekkles).</p>
+        <p>* Fixed boss list for Throne of Thunder.</p>
+        <p>* (MoP) Added Eternal Flame for "Healers".</p>
+        <br/>
+
+        <h1>r271-release (2025-12-08 12:00 GMT+8)</h1>
+        <p>* Fixed a OmniCD related issue.</p>
+        <p>* (Classic) The "Debuffs" indicator no longer shows WeakenedSoul for non-priests.</p>
+        <br/>
+
+        <h1>r270-release (2025-12-04 10:00 GMT+8)</h1>
+        <p>* Updated toc version.</p>
+        <p>* Fixed display of the separate pet frame in arenas.</p>
+        <br/>
+
+        <h1>r269-release (2025-11-04 10:29 GMT+8)</h1>
+        <p>* (CN MoP) Fixed Monk-related issues.</p>
+        <br/>
+
+        <h1>r268-release (2025-10-30 11:15 GMT+8)</h1>
+        <p>* Fixed spell rank issue for the English client.</p>
+        <p>* Fixed potential nil error in Actions indicator.</p>
+        <p>* Updated locales.</p>
+        <br/>
+
+        <h1>r267-release (Oct 27, 2025, 17:50 GMT+8)</h1>
+        <p>* Fixed specID check (layout switch) on auto spec change.</p>
+        <p>* Fixed spell ranks for click-castings.</p>
+        <p>+ Added esMX (thanks maylisdalan).</p>
+        <p>* Updated locales, fixed missing phrases.</p>
+        <br/>
+
         <h1>r266-release (Oct 14, 2025, 14:25 GMT+8)</h1>
         <p>* Delayed creation of some indicators to avoid potential issues.</p>
         <br/>