+++ AddOns/Cell/Modules/ClickCastings/ClickCastings.lua
@@ -3,8 +3,6 @@ local L = Cell.L
 local F = Cell.funcs
 local P = Cell.pixelPerfectFuncs
 
-local RANK = _G.TRADESKILL_RANK_HEADER:gsub(" ", ""):gsub("%%d", "")
-
 local clickCastingsTab = Cell.CreateFrame("CellOptionsFrame_ClickCastingsTab", Cell.frames.optionsFrame, nil, nil, true)
 Cell.frames.clickCastingsTab = clickCastingsTab
 clickCastingsTab:SetAllPoints(Cell.frames.optionsFrame)
@@ -525,7 +523,7 @@ local function ApplyClickCastings(b)
             spellName = spellName or ""
 
             if rank then
-                spellName = spellName.."(".. RANK .. " " .. rank ..")"
+                spellName = spellName .. F.GetRankSuffix(rank)
             end
 
             --! NOTE: fix Primordial Wave
@@ -1154,12 +1152,22 @@ local function ShowActionsMenu(index, b)
             end,
         })
 
-        if (Cell.isVanilla or Cell.isWrath or Cell.isCata) and Cell.vars.playerClass == "WARLOCK" then
+        if (Cell.isVanilla or Cell.isTBC or Cell.isWrath or Cell.isCata) and Cell.vars.playerClass == "WARLOCK" then
+            local soulstoneID
+            if Cell.isVanilla then
+                soulstoneID = 16896
+            elseif Cell.isTBC then
+                soulstoneID = 22116
+            else -- wrath & cata
+                soulstoneID = 36895
+            end
+
             tinsert(items, {
                 ["text"] = F.GetSpellInfo(20707),
                 ["onClick"] = function()
                     changed[index] = changed[index] or {b}
-                    local macrotext = "/stopcasting\n/target mouseover\n/use item:36895\n/targetlasttarget"
+
+                    local macrotext = "/stopcasting\n/target mouseover\n/use item:"..soulstoneID.."\n/targetlasttarget"
                     if b.bindAction ~= macrotext then
                         changed[index]["bindAction"] = macrotext
                         b.actionGrid:SetText(macrotext)
@@ -1349,11 +1357,11 @@ local function ShowActionsMenu(index, b)
                 end
             }
 
-            if t[5] and t[5] > 1 then
+            if t[5] and t[5] >= 1 then
                 spellItem.children = {}
                 for i = 1, t[5] do
                     tinsert(spellItem.children, {
-                        ["text"] = RANK .. " " .. i,
+                        ["text"] = i,
                         ["onClick"] = function()
                             changed[index] = changed[index] or {b}
                             if b.bindAction ~= t[4]..":"..i then
@@ -1377,7 +1385,7 @@ local function ShowActionsMenu(index, b)
     end
 
     menu:SetItems(items, 15)
-    menu:SetWidths(b.actionGrid:GetWidth(), 70)
+    menu:SetWidths(b.actionGrid:GetWidth(), 35)
     P.ClearPoints(menu)
     P.Point(menu, "TOPLEFT", b.actionGrid, "BOTTOMLEFT", 0, -1)
     menu:ShowMenu()
@@ -1396,7 +1404,7 @@ local function UpdateCurrentText(isCommon)
     else
         if Cell.isRetail or Cell.isMists then
             listPane:SetTitle(L["Current Profile"]..": ".."|T"..Cell.vars.playerSpecIcon..":12:12:0:1:12:12:1:11:1:11|t "..Cell.vars.playerSpecName)
-        elseif Cell.isCata or Cell.isWrath or Cell.isVanilla then
+        elseif Cell.isCata or Cell.isWrath or Cell.isTBC or Cell.isVanilla then
             local name, icon = F.GetActiveTalentInfo()
             listPane:SetTitle(L["Current Profile"]..": ".."|T"..icon..":12:12:0:1:12:12:1:11:1:11|t "..name)
         end